requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.7
langchain-huggingface>=0.3.0
langchain-ollama>=0.3.0
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.0
transformers>=4.35.0
torch>=2.2.0
Pillow>=10.0.0
pdf2image>=1.16.0
docx2pdf>=0.1.8
pytesseract>=0.3.10
requests>=2.31.0